+++ module-audio/Audio/Operation/Operation.cpp
@@ -10,12 +10,14 @@
 #include "RecorderOperation.hpp"
 #include "RouterOperation.hpp"
 
-#include "bsp/audio/bsp_audio.hpp"
 #include "Audio/decoder/decoder.hpp"
+
+#include <bsp/audio/bsp_audio.hpp>
+#include <module-bsp/board/rt1051/bsp/audio/RT1051BluetoothAudio.hpp>
 
 namespace audio
 {
-    std::optional<std::unique_ptr<Operation>> Operation::Create(
+    std::unique_ptr<Operation> Operation::Create(
         Operation::Type t,
         const char *fileName,
         const audio::PlaybackType &playbackType,
@@ -48,16 +50,46 @@ namespace audio
         }
     }
 
-    std::optional<std::shared_ptr<Profile>> Operation::GetProfile(const Profile::Type type)
+    std::shared_ptr<Profile> Operation::GetProfile(const Profile::Type type)
     {
-        auto ret = std::find_if(
-            availableProfiles.begin(), availableProfiles.end(), [type](const auto &w) { return w->GetType() == type; });
-        if (ret == availableProfiles.end()) {
-            return std::nullopt;
+        auto ret = std::find_if(supportedProfiles.begin(), supportedProfiles.end(), [type](const auto &w) {
+            return w.isAvailable == true && w.profile->GetType() == type;
+        });
+        if (ret == supportedProfiles.end()) {
+            return nullptr;
         }
         else {
-            return *ret;
+            return ret->profile;
         }
     }
 
+    void Operation::SetBluetoothStreamData(std::shared_ptr<BluetoothStreamData> data)
+    {
+        if (auto device = dynamic_cast<bsp::RT1051BluetoothAudio *>(audioDevice.get()); device != nullptr) {
+            GetProfile()->SetSampleRate(data->metadata.sampleRate);
+            device->sourceQueue = data->out;
+            device->metadata    = data->metadata;
+            LOG_INFO("Queue and metadata set!");
+        }
+    }
+
+    audio::RetCode Operation::SwitchToPriorityProfile()
+    {
+        for (auto &p : supportedProfiles) {
+            if (p.isAvailable == true) {
+                return SwitchProfile(p.profile->GetType());
+            }
+        }
+        return audio::RetCode::Failed;
+    }
+
+    void Operation::SetProfileAvailability(std::vector<Profile::Type> profiles, bool available)
+    {
+        for (auto &p : supportedProfiles) {
+            if (auto shouldSet = (std::find(profiles.begin(), profiles.end(), p.profile->GetType()) != profiles.end());
+                shouldSet) {
+                p.isAvailable = available;
+            }
+        }
+    }
 } // namespace audio