+++ module-bsp/board/rt1051/bsp/audio/CodecMAX98090.hpp
@@ -6,6 +6,7 @@
 
 #include "Codec.hpp"
 #include "drivers/i2c/DriverI2C.hpp"
+#include "bsp/audio/bsp_audio.hpp"
 
 class CodecParamsMAX98090 : public CodecParams
 {
@@ -31,22 +32,6 @@ class CodecParamsMAX98090 : public CodecParams
         Rate32KHz  = 32000,
         Rate96KHz  = 96000,
         Invalid
-    };
-
-    enum class OutputPath
-    {
-        Headphones,
-        HeadphonesMono,
-        Earspeaker,
-        Loudspeaker,
-        None
-    };
-
-    enum class InputPath
-    {
-        Headphones,
-        Microphone,
-        None
     };
 
     enum class MonoStereo
@@ -101,8 +86,8 @@ class CodecParamsMAX98090 : public CodecParams
     bool muteEnable       = false;
     bool resetEnable      = false;
     bool micBiasEnable    = false;
-    InputPath inputPath   = InputPath ::None;
-    OutputPath outputPath = OutputPath ::None;
+    bsp::AudioDevice::InputPath inputPath   = bsp::AudioDevice::InputPath::None;
+    bsp::AudioDevice::OutputPath outputPath = bsp::AudioDevice::OutputPath::None;
     SampleRate sampleRate = SampleRate ::Rate44K1Hz;
 };
 
@@ -132,8 +117,8 @@ class CodecMAX98090 : public Codec
     CodecRetCode SetOutputVolume(const float vol);
     CodecRetCode SetInputGain(const float gain);
     CodecRetCode SetMute(const bool enable);
-    CodecRetCode SetInputPath(const CodecParamsMAX98090::InputPath path);
-    CodecRetCode SetOutputPath(const CodecParamsMAX98090::OutputPath path);
+    CodecRetCode SetInputPath(const bsp::AudioDevice::InputPath path);
+    CodecRetCode SetOutputPath(const bsp::AudioDevice::OutputPath path);
     CodecRetCode MicBias(const bool enable);
     CodecRetCode WriteFilterCoeff(const float coeff, const uint8_t basereg);
     CodecRetCode Reset();