+++ module-services/service-audio/service-audio/AudioMessage.hpp
@@ -232,7 +232,8 @@ class AudioEventRequest : public AudioMessage
     explicit AudioEventRequest(std::shared_ptr<audio::Event> evt) : evt(std::move(evt))
     {}
 
-    explicit AudioEventRequest(audio::EventType eType) : evt(std::make_shared<audio::Event>(eType))
+    explicit AudioEventRequest(audio::EventType eType, audio::Event::DeviceState state)
+        : evt(std::make_shared<audio::Event>(eType, state))
     {}
 
     std::shared_ptr<audio::Event> getEvent()